requirements.txt
web3==6.12.0
kafka-python==2.0.2
python-dotenv==1.0.1
